+++ src/main/java/com/example/burrowwebapp/models/User.java
@@ -2,11 +2,10 @@ package com.example.burrowwebapp.models;
 
 import com.sun.istack.NotNull;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-
 import javax.persistence.Entity;
 
 @Entity
-public class User {
+public class User extends AbstractEntity {
 
     private static final BCryptPasswordEncoder encoder = new BCryptPasswordEncoder();
 
@@ -20,11 +19,15 @@ public class User {
 
     public User(String username, String password) {
         this.username = username;
-        this.pwHash = password;
+        this.pwHash = encoder.encode(password);
     }
 
     public String getUsername() {
         return username;
+    }
+
+    public boolean isMatchingPassword(String password) {
+        return encoder.matches(password, pwHash);
     }
 
 }
